// frontend-v3/src/app/core/api/models/cliente-input.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ClienteEnderecoInput } from '../models/cliente-endereco-input';

/**
 * Dados do cliente
 */
export interface ClienteInput {

  /**
   * CPF ou CNPJ do cliente
   */
  documento?: string;

  /**
   * Lista de enderecos do cliente
   */
  enderecos: Array<ClienteEnderecoInput>;

  /**
   * Nome do cliente
   */
  nome: string;

  /**
   * Telefone principal
   */
  tel_1?: string;

  /**
   * Telefone secundario
   */
  tel_2?: string;

  /**
   * Telefone adicional
   */
  tel_3?: string;
}