// frontend-v3/src/app/core/api/models/cliente-output.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ClienteEnderecoOutput } from '../models/cliente-endereco-output';

/**
 * Dados de saida do cliente
 */
export interface ClienteOutput {

  /**
   * CPF ou CNPJ do cliente
   */
  documento?: string;

  /**
   * Lista de enderecos do cliente
   */
  enderecos: Array<ClienteEnderecoOutput>;

  /**
   * ID do cliente
   */
  id: number;

  /**
   * Nome do cliente
   */
  nome: string;

  /**
   * Telefone principal
   */
  tel_1?: string;

  /**
   * Telefone secundario
   */
  tel_2?: string;

  /**
   * Telefone adicional
   */
  tel_3?: string;
}